added example temporal graph - no spatial axes
